Allow storage policy to determine whether k-configurations are unpacked from the twopf database when a threepf configuration is rejected
- in some cases we want rejected k-configurations of this kind to be retained, for example if we are trying to build a regular grid in k or alpha. Then we want to assign serial numbers once, in order, and keep them, so it's a bad idea to unpack them from the database

- new return type for storage policy allows a configuration to be accepted, rejected retaining 2pf configurations, or rejected with removal of 2pf configurations

+++ CppTransport/transport-runtime-api/tasks/configuration-database/threepf_config_database.h
@@ -495,7 +495,8 @@ namespace transport
         if(k3_stored) config.k3_serial = (*rec)->serial;
         else          config.k3_serial = twopf_db.add_record(config.k3_conventional);
 
-        if(policy(config))   // policy confirms that this configuration should be retained
+        storage_outcome result = policy(config);
+        if(result == storage_outcome::accept)   // policy confirms that this configuration should be retained
           {
             if(config.kt_conventional > this->kmax_conventional) this->kmax_conventional = config.kt_conventional;
             if(config.kt_conventional < this->kmin_conventional) this->kmin_conventional = config.kt_conventional;
@@ -514,10 +515,14 @@ namespace transport
 
         // policy declined to store this configuration
 
-        // unwind any twopf configurations added
-        if(!k1_stored) twopf_db.delete_record(config.k1_serial);
-        if(!k2_stored) twopf_db.delete_record(config.k2_serial);
-        if(!k3_stored) twopf_db.delete_record(config.k3_serial);
+        if(result == storage_outcome::reject_remove)
+          {
+            // unwind any twopf configurations added
+            if(!k1_stored) twopf_db.delete_record(config.k1_serial);
+            if(!k2_stored) twopf_db.delete_record(config.k2_serial);
+            if(!k3_stored) twopf_db.delete_record(config.k3_serial);
+            twopf_db.rebuild_cache();
+          }
 
         return boost::optional<unsigned int>();
       }
